refactor: extract PendingTranscription types from TranscriptSession.swift (#651)
Bundles both pending-transcription types (PendingTranscriptionFailureReason enum
+ PendingTranscription struct) into a new PendingTranscription.swift.

Both types are part of the same intent domain (pending-transcription lifecycle
metadata) — bundling them into one file is intent-first coherent.

Byte-identical moves; SHAs verified against origin/main:
- PendingTranscriptionFailureReason enum (lines 3-115):
  9fedcb1ef3304f8f011e98a30b9fc008dda7d451565958d27c04c38966519f87
- PendingTranscription struct (lines 117-146):
  a4b5303df88d66dac82a23845686fad16cd24a292fe1c23902db6053d26a571c

Reduces TranscriptSession.swift from 686 to 541 lines.

Tests: 667 (unchanged).
